Move _get_logging_config into sprockets.http.
