+++ sprockets/http/__init__.py
@@ -62,10 +62,58 @@ def run(create_application, settings=None):
     debug_mode = bool(app_settings.get('debug',
                                        int(os.environ.get('DEBUG', 0)) != 0))
     app_settings['debug'] = debug_mode
-    logging.config.dictConfig(runner._get_logging_config(debug_mode))
+    logging.config.dictConfig(_get_logging_config(debug_mode))
 
     port_number = int(app_settings.pop('port', os.environ.get('PORT', 8000)))
     num_procs = int(app_settings.pop('number_of_procs', '0'))
     server = runner.Runner(create_application(**app_settings))
 
     server.run(port_number, num_procs)
+
+
+def _get_logging_config(debug):
+    if debug:
+        return {
+            'version': 1,
+            'disable_existing_loggers': False,
+            'incremental': False,
+            'formatters': {
+                'debug': {
+                    'format': ('[%(asctime)s] %(levelname)-8s %(process)-6s '
+                               '%(name)s: %(message)s.')
+                },
+            },
+            'handlers': {
+                'debug-console': {
+                    'class': 'logging.StreamHandler',
+                    'stream': 'ext://sys.stdout',
+                    'level': 'DEBUG',
+                    'formatter': 'debug',
+                },
+            },
+            'root': {
+                'level': 'DEBUG',
+                'handlers': ['debug-console'],
+            }
+        }
+    else:
+        return {
+            'version': 1,
+            'disable_existing_loggers': False,
+            'incremental': False,
+            'formatters': {
+                'json': {
+                    '()': 'sprockets.logging.JSONRequestFormatter',
+                },
+            },
+            'handlers': {
+                'console': {
+                    'class': 'logging.StreamHandler',
+                    'formatter': 'json',
+                },
+            },
+            'root': {
+                'level': 'INFO',
+                'handlers': ['console'],
+            }
+        }
